Move search functionality to `search.js`
